Add files via upload
Aim of this code is to perform binary classification of various news articles available online with the help of concepts pertaining to Artificial Intelligence, Natural Language Processing and Machine Learning. It aims to provide the user with the ability to classify the news as fake or real. It is used to propose machine learning approach for automated classification of news articles. This project explores different textual properties that can be used to distinguish fake contents from real. By using those properties, will train a machine learning algorithms and evaluate their performance datasets.